move method closePolis to CustomerClaimService
