+++ src/main/java/com/app/smartdrive/api/controllers/customer/CustomerRequestController.java
@@ -156,7 +156,7 @@ public class CustomerRequestController {
 
     @PutMapping("/close")
     public ResponseEntity<CustomerResponseDTO> requestClosePolis(@RequestBody CloseRequestDTO closeRequestDTO){
-        CustomerResponseDTO customerResponseDTO = this.customerRequestService.closePolis(closeRequestDTO);
+        CustomerResponseDTO customerResponseDTO = this.customerClaimService.closePolis(closeRequestDTO);
 
         return ResponseEntity.status(HttpStatus.OK).body(customerResponseDTO);
     }
