+++ src/main/java/com/app/smartdrive/api/services/customer/CustomerRequestService.java
@@ -37,8 +37,6 @@ public interface CustomerRequestService {
 
     public void changeRequestTypeToClaim(CustomerRequestTypeDTO customerRequestTypeDTO);
 
-    public CustomerResponseDTO closePolis(CloseRequestDTO closeRequestDTO);
-
     public void changeRequestTypeToClose(CustomerRequestTypeDTO customerRequestTypeDTO);
 
 }
